Move getDateStartAndEndTimes to TimeUtils
TimeUtils as a location is much more logical than DataUtils.
